Workshop/ETL/requirements.txt
pandas>=1.5.0
google-cloud-bigquery>=3.0.0
scikit-learn>=1.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
